Export KPIECE and RRT solvers
