--- exotations/solvers/exotica_control_rrt_solver/exotica_plugins.xml
@@ -1,5 +1,0 @@
-<library path="lib/libexotica_control_rrt_solver">
-  <class name="exotica/ControlRRTSolver" type="exotica::ControlRRTSolver" base_class_type="exotica::MotionSolver">
-    <description>Kinodynamic Control-RRT Solver from OMPL</description>
-  </class>
-</library>
