+++ exotations/solvers/exotica_ompl_control_solver/include/exotica_ompl_control_solver/ompl_native_solvers.h
@@ -1,0 +1,58 @@
+//
+// Copyright (c) 2019, University of Edinburgh
+// All rights reserved.
+//
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions are met:
+//
+//  * Redistributions of source code must retain the above copyright notice,
+//    this list of conditions and the following disclaimer.
+//  * Redistributions in binary form must reproduce the above copyright
+//    notice, this list of conditions and the following disclaimer in the
+//    documentation and/or other materials provided with the distribution.
+//  * Neither the name of  nor the names of its contributors may be used to
+//    endorse or promote products derived from this software without specific
+//    prior written permission.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+// ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
+// LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+// CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+// SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+// INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+// CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+// ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+// POSSIBILITY OF SUCH DAMAGE.
+//
+
+#ifndef EXOTICA_OMPL_CONTROL_SOLVER_OMPL_NATIVE_SOLVERS_H_
+#define EXOTICA_OMPL_CONTROL_SOLVER_OMPL_NATIVE_SOLVERS_H_
+
+#include <exotica_ompl_control_solver/ompl_control_solver.h>
+#include <ompl/control/planners/kpiece/KPIECE1.h>
+#include <ompl/control/planners/rrt/RRT.h>
+
+#include <exotica_ompl_control_solver/control_kpiece_solver_initializer.h>
+#include <exotica_ompl_control_solver/control_rrt_solver_initializer.h>
+
+namespace exotica
+{
+class ControlRRTSolver : public OMPLControlSolver<ompl::control::RRT>, Instantiable<ControlRRTSolverInitializer>
+{
+public:
+    ControlRRTSolver();
+    void Instantiate(const ControlRRTSolverInitializer& init) override;
+};
+
+class ControlKPIECESolver : public OMPLControlSolver<ompl::control::KPIECE1>, Instantiable<ControlKPIECESolverInitializer>
+{
+public:
+    ControlKPIECESolver();
+    void Instantiate(const ControlKPIECESolverInitializer& init) override;
+};
+
+}  // namespace exotica
+
+#endif  // EXOTICA_OMPL_CONTROL_SOLVER_OMPL_NATIVE_SOLVERS_H_
