+++ exotations/solvers/exotica_ompl_control_solver/init/ompl_control_solver.in
@@ -1,4 +1,4 @@
-class ControlRRTSolver
+class OMPLControlSolver
 
 extend <exotica_core/motion_solver>
 Required Eigen::VectorXd StateLimits;
