+++ exotations/solvers/exotica_ompl_control_solver/src/ompl_native_solvers.cpp
@@ -1,0 +1,55 @@
+//
+// Copyright (c) 2018, University of Edinburgh
+// All rights reserved.
+//
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions are met:
+//
+//  * Redistributions of source code must retain the above copyright notice,
+//    this list of conditions and the following disclaimer.
+//  * Redistributions in binary form must reproduce the above copyright
+//    notice, this list of conditions and the following disclaimer in the
+//    documentation and/or other materials provided with the distribution.
+//  * Neither the name of  nor the names of its contributors may be used to
+//    endorse or promote products derived from this software without specific
+//    prior written permission.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+// ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
+// LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+// CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+// SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+// INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+// CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+// ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+// POSSIBILITY OF SUCH DAMAGE.
+//
+
+#include <exotica_ompl_control_solver/ompl_native_solvers.h>
+
+REGISTER_MOTIONSOLVER_TYPE("ControlRRTSolver", exotica::ControlRRTSolver)
+REGISTER_MOTIONSOLVER_TYPE("ControlKPIECESolver", exotica::ControlKPIECESolver)
+
+namespace exotica
+{
+// ControlRRTSolver
+ControlRRTSolver::ControlRRTSolver() = default;
+
+void ControlRRTSolver::Instantiate(const ControlRRTSolverInitializer &init)
+{
+    init_ = OMPLControlSolverInitializer(ControlRRTSolverInitializer(init));
+    algorithm_ = "ControlRRTSolver";
+}
+
+// ControlKPIECESolver
+ControlKPIECESolver::ControlKPIECESolver() = default;
+
+void ControlKPIECESolver::Instantiate(const ControlKPIECESolverInitializer &init)
+{
+    init_ = OMPLControlSolverInitializer(ControlKPIECESolverInitializer(init));
+    algorithm_ = "ControlKPIECESolver";
+}
+
+}  // namespace exotica
