+++ exotica_examples/resources/configs/dynamic_time_indexed/15_rrt_quadrotor.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" ?>
 <DynamicTimeIndexedProblemConfig>
-    <ControlRRTSolver Name="OMPLKdSolver">
+    <ControlRRTSolver Name="ControlRRTSolver">
         <Debug>1</Debug>
         <StateLimits>3 3 3 1 1 1 3 3 3 1 1 1</StateLimits>
         <MaxIterationTime>10</MaxIterationTime>
