+++ exotica_examples/scripts/example_dynamic_time_indexed
@@ -31,6 +31,7 @@ configs = [
     '{exotica_examples}/resources/configs/dynamic_time_indexed/13_control_limited_ddp_quadrotor.xml',
     '{exotica_examples}/resources/configs/dynamic_time_indexed/14_rrt_cartpole.xml',
     '{exotica_examples}/resources/configs/dynamic_time_indexed/15_rrt_quadrotor.xml',
+    '{exotica_examples}/resources/configs/dynamic_time_indexed/16_kpiece_cartpole.xml',
 ]
 
 problem_idx = int(sys.argv[1]) - 1
